# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PixelForge Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface GenerateSpriteInput {
  /** Text description of the desired pixel art */
  prompt: string;
  /** Art style preset: blasphemous, retro8bit, 16bit, isometric, gothic */
  style?: string;
  /** Target pixel art width */
  width?: number;
  /** Target pixel art height */
  height?: number;
  /** Number of animation frames to generate (1 = static) */
  frameCount?: number;
  /** Optional base64 reference image for style consistency */
  referenceImage?: string;
}

export interface GenerateSpriteOutput {
  /** Array of base64-encoded PNG frames (1024x1024 source images) */
  frames: string[];
  /** The enhanced prompt that was used */
  prompt: string;
  style: string;
}

export interface EnhanceImageInput {
  /** Base64-encoded source image to convert to pixel art */
  imageBase64: string;
  /** Optional style guidance */
  prompt?: string;
  style?: string;
  width?: number;
  height?: number;
}

export interface ProjectSummary {
  id: string;
  name: string;
  width: number;
  height: number;
  mode: string;
  thumbnail?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface ProjectFull {
  id: string;
  name: string;
  width: number;
  height: number;
  mode: string;
  /** Full ProjectData JSON */
  data: unknown;
  thumbnail?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  id: string;
  name: string;
  width: number;
  height: number;
  mode?: string;
  /** Full ProjectData JSON */
  data: unknown;
  thumbnail?: string | null;
  tags?: string[];
}

export type ListProjectsParams = {
mode?: string;
search?: string;
};