# lib/api-zod/src/generated/types/enhanceImageInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PixelForge Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface EnhanceImageInput {
  /** Base64-encoded source image to convert to pixel art */
  imageBase64: string;
  /** Optional style guidance */
  prompt?: string;
  style?: string;
  width?: number;
  height?: number;
}