# lib/api-zod/src/generated/types/generateSpriteInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PixelForge Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateSpriteInput {
  /** Text description of the desired pixel art */
  prompt: string;
  /** Art style preset: blasphemous, retro8bit, 16bit, isometric, gothic */
  style?: string;
  /** Target pixel art width */
  width?: number;
  /** Target pixel art height */
  height?: number;
  /** Number of animation frames to generate (1 = static) */
  frameCount?: number;
  /** Optional base64 reference image for style consistency */
  referenceImage?: string;
}