# lib/api-zod/src/generated/types/generateSpriteOutput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PixelForge Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateSpriteOutput {
  /** Array of base64-encoded PNG frames (1024x1024 source images) */
  frames: string[];
  /** The enhanced prompt that was used */
  prompt: string;
  style: string;
}