# lib/api-zod/src/generated/types/projectFull.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PixelForge Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectFull {
  id: string;
  name: string;
  width: number;
  height: number;
  mode: string;
  /** Full ProjectData JSON */
  data: unknown;
  thumbnail?: string | null;
  tags: string[];
  createdAt: Date;
  updatedAt: Date;
}