# lib/api-zod/src/generated/types/projectSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PixelForge Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectSummary {
  id: string;
  name: string;
  width: number;
  height: number;
  mode: string;
  thumbnail?: string | null;
  tags: string[];
  createdAt: Date;
  updatedAt: Date;
}